requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.4
djangorestframework==3.15.2
et_xmlfile==2.0.0
idna==3.10
numpy==2.2.1
openpyxl==3.1.5
pandas==2.2.3
psycopg2-binary==2.9.10
pyaes==1.6.1
pyasn1==0.6.1
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
rsa==4.9
six==1.17.0
sqlparse==0.5.3
Telethon==1.38.1
tzdata==2024.2
urllib3==2.3.0
xlwings==0.33.6
